fix(e2e): move MIGRATION_BACKUP_PREFIX to migration.const
