+++ src/app/op-log/persistence/migration.const.ts
@@ -1,0 +1,5 @@
+/**
+ * Constant for migration backup file naming.
+ * Used across both Angular app and E2E tests.
+ */
+export const MIGRATION_BACKUP_PREFIX = 'sp-pre-migration-backup';
